pass the raw params for signature verification.
The HTTP Redirect binding demands that the unescaped values are used for
signature verification. Rails unescapes the params before passing it to
the application code which is useful in most cases except for this one.

https://docs.oasis-open.org/security/saml/v2.0/saml-bindings-2.0-os.pdf

The relying party MUST therefore perform the verification step using the
original URL-encoded
values it received on the query string. It is not sufficient to
re-encode the parameters after they have been
processed by software because the resulting encoding may not match the
signer's encoding.